move smart inflation settings from DeveloperPreferences to CommCarePreferences, making compatible with setting up from HQ
